requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.38
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pypdf==5.3.0
pillow==11.1.0
pytesseract==0.3.13
openai==1.63.2
scikit-learn==1.6.1
numpy==2.2.3
